+++ .gitignore
@@ -1,2 +1,3 @@
 dst/
 testflow.txt
+node_modules
