+++ ruby/.scaffold/.babelrc
@@ -1,0 +1,6 @@
+{
+  "presets": [
+    ["env", { "modules": false }],
+    "stage-3"
+  ]
+}
