+++ ruby/.scaffold/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <title>atwt</title>
+  </head>
+  <body>
+    <div id="app">
+    </div>
+    <script src="/dist/build.js"></script>
+  </body>
+</html>
