+++ ruby/.scaffold/src/routes.js
@@ -1,0 +1,7 @@
+
+export default [
+  {
+    path: "/",
+    component: Entry
+  },
+]
